21.04.26 11:41 강현 푸시 (완료)   network function 함수 추가   user의 isBusiness 상태를 바꿔주는 함수
** user와 store 결합 진행중 **

(완료)
home screen에서 userModelState 연결

(완료)
menu 부분
 - userModelState 연결 및 isBusiness Ui 구현
 - 액티비티, 이용내역(x), 리뷰관리, 관심목록 들어가는 것 구현
 - business transfer page 추가

(예정)
store 등록, favorites, reviews 불러오기 연결
각 페이지 구현

(진행예정)
 location 설정 및 pin 포인트 지정 가능하도록

+++ lib/model/user_model.dart
@@ -9,7 +9,7 @@ class UserModel {
   final String profileImg;
   final bool isBusiness;
   final List<dynamic> userLocations; // dynamic : Map 형태  {'title': '집', 'address': '대전 유성구 대학로 291', 'detail': 'w5-2 스타트업빌리지 204호', 'lat': 36.3693545, 'lon':127.355723}
-  final double range;
+  final int range;  // 5000m , 10000m 요런식으로 사용할 예정
   final List<dynamic> activities;
   final List<dynamic> favorites;
   final List<dynamic> myReviews;
